refactor: make screener question controller a mixin
